% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{build_prior_from_specs}
\alias{build_prior_from_specs}
\title{Build phase_prior object from parsed specifications}
\usage{
build_prior_from_specs(specs)
}
\description{
Build phase_prior object from parsed specifications
}
\keyword{internal}
