% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{build_re_z_matrix}
\alias{build_re_z_matrix}
\title{Build Z matrix (RE design matrix) for a term with slopes}
\usage{
build_re_z_matrix(data, term)
}
\arguments{
\item{data}{Data frame}

\item{term}{RE term specification}
}
\value{
Matrix with n_obs rows and n_coefs columns
}
\description{
Build Z matrix (RE design matrix) for a term with slopes
}
\keyword{internal}
