% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{cauchy}
\alias{cauchy}
\title{Cauchy prior distribution}
\usage{
cauchy(location = 0, scale = 1)
}
\arguments{
\item{location}{Location parameter}

\item{scale}{Scale parameter (must be positive)}
}
\value{
A prior specification object
}
\description{
Cauchy prior distribution
}
\examples{
prior(beta ~ cauchy(0, 2.5))

}
