% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_hmc.R
\name{combine_diagnostics}
\alias{combine_diagnostics}
\title{Combine diagnostics from multiple chains}
\usage{
combine_diagnostics(draws_list)
}
\description{
Combine diagnostics from multiple chains
}
\keyword{internal}
