% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{compare_waic}
\alias{compare_waic}
\title{Compare models using WAIC}
\usage{
compare_waic(..., model_names = NULL)
}
\arguments{
\item{...}{\code{phase_fit} objects to compare}

\item{model_names}{Optional names for the models}
}
\value{
A data frame with WAIC comparisons
}
\description{
Compare multiple fitted phase models using WAIC.
}
\examples{
\dontrun{
compare_waic(fit1, fit2, model_names = c("Simple", "Complex"))
}

}
