% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{convert_to_phase_fit}
\alias{convert_to_phase_fit}
\title{Convert raw fit to phase_fit object}
\usage{
convert_to_phase_fit(
  fit_raw,
  model,
  data,
  backend,
  chains,
  iter,
  warmup,
  stan_data
)
}
\description{
Convert raw fit to phase_fit object
}
\keyword{internal}
