% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{create_nested_interactions}
\alias{create_nested_interactions}
\title{Create interaction variables for nested RE}
\usage{
create_nested_interactions(data, interactions)
}
\arguments{
\item{data}{Data frame}

\item{interactions}{Character vector of interaction terms (e.g., "site:plot")}
}
\value{
Modified data frame with new columns
}
\description{
Creates columns like "site:plot" from "site" and "plot" columns.
}
\keyword{internal}
