% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{estimate_pareto_k}
\alias{estimate_pareto_k}
\title{Estimate Pareto shape parameter}
\usage{
estimate_pareto_k(x)
}
\arguments{
\item{x}{Exceedances (positive values)}
}
\value{
Estimated shape parameter k
}
\description{
Uses probability-weighted moments estimator for the generalized Pareto
distribution shape parameter.
}
\keyword{internal}
