% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example11.R
\name{example11}
\alias{example11}
\title{Example ODE System 11}
\usage{
example11(t, y, parameters)
}
\arguments{
\item{t}{The value of \ifelse{html}{\out{<i>t</i>}}{\eqn{t}}, the independent
variable, to evaluate the derivative at. Should be a single number.}

\item{y}{The values of \ifelse{html}{\out{<i>x</i>}}{\eqn{x}} and 
\ifelse{html}{\out{<i>y</i>}}{\eqn{y}}, the dependent
variables, to evaluate the derivative at. Should be a vector of length two.}

\item{parameters}{The values of the parameters of the system. Not required
here.}
}
\value{
Returns a list \code{dy} containing the values of the two derivatives
at \ifelse{html}{\out{(<i>t</i>, <i>x</i>, <i>y</i>)}}{\eqn{(t, x, y)}}.
}
\description{
The derivative function of an example two-dimensional autonomous ODE system.
}
\details{
\code{example11} evaluates the derivatives of the following coupled ODE system
at the point \ifelse{html}{\out{(<i>t</i>, <i>x</i>, <i>y</i>)}}{
\eqn{(t, x, y)}}:

\ifelse{html}{\out{<center><i>dx</i>/<i>dt</i> = <i>x</i>(3 - <i>x</i> - 2<i>y</i>),
<i>dy</i>/<i>dt</i> = -<i>y</i>(2 - <i>x</i> - <i>y</i>).</center>}}{\deqn{dx/dt = x(3 - x - 2y), dy/dt = -y(2 - x - y).}}

Its format is designed to be compatible with \code{\link[deSolve]{ode}} from
the \code{\link[deSolve]{deSolve}} package.
}
\examples{
# Plot the velocity field, nullclines and several trajectories
example11.flowField   <- flowField(example11,
                                   xlim = c(-5, 5),
                                   ylim = c(-5, 5),
                                   points = 21,
                                   add = FALSE)
y0                    <- matrix(c(4, 4, -1, -1,
                                  -2, 1, 1, -1), 4, 2,
                                byrow = TRUE)
example11.nullclines  <- nullclines(example11,
                                    xlim = c(-5, 5),
                                    ylim = c(-5, 5),
                                    points = 200)
example11.trajectory  <- trajectory(example11,
                                    y0 = y0,
                                    tlim = c(0, 10))
example11.manifolds   <- drawManifolds(example11,
                                       y0 = c(1, 1),
                                       tend = 1000,
                                       col = c("green", "red"),
                                       add.legend = TRUE)
# Determine the stability of the equilibrium points
example11.stability.1 <- stability(example11, ystar = c(0, 0))
example11.stability.2 <- stability(example11, ystar = c(0, 2))
example11.stability.3 <- stability(example11, ystar = c(1, 1))
example11.stability.4 <- stability(example11, ystar = c(3, 0))
}
\seealso{
\code{\link[deSolve]{ode}}
}
\author{
Michael J. Grayling
}
