% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example13.R
\name{example13}
\alias{example13}
\title{Example ODE System Number Thirteen}
\usage{
example13(t, y, parameters)
}
\arguments{
\item{t}{Value of t, the independent variable, to evaluate the derivative
at. Should be a single number.}

\item{y}{Values of x and y, the dependent variables, to evaluate the
derivative at. Should be a vector of length 2.}

\item{parameters}{Values of the parameters of the system. Not required here.}
}
\value{
Returns a list dy containing the values of the two derivatives at
(t, x, y).
}
\description{
The derivative function of an example two dimensional autonomous ODE system.
}
\details{
Evaluates the derivative of the following coupled ODE system at the point
(t, x, y):

dx/dt = 2 - x^2 - y^2, dy/dt = x^2 - y^2.

Format is designed to be compatible with ode from the deSolve package.
}
\seealso{
\code{\link{ode}}
}
\author{
Michael J. Grayling
}
