% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example4.R
\name{example4}
\alias{example4}
\title{Example ODE System Number Four}
\usage{
example4(t, y, parameters)
}
\arguments{
\item{t}{Value of t, the independent variable, to evaluate the derivative
at. Should be a single number.}

\item{y}{Values of x and y, the dependent variables, to evaluate the
derivative at. Should be a vector of length 2.}

\item{parameters}{Values of the parameters of the system. Not required here.}
}
\value{
Returns a list dy containing the values of the two derivatives at
(t, x, y).
}
\description{
The derivative function of an example two dimensional autonomous ODE system.
}
\details{
Evaluates the derivatives of the following coupled ODE system at the point
(t, x, y):

dx/dt = -x, dy/dt = 4*x.

Format is designed to be compatible with ode from the deSolve package.
}
\examples{
# Plot the velocity field, nullclines and several trajectories.
example4.flowField  <- flowField(example4, xlim = c(-3, 3), ylim = c(-5, 5),
                                 points = 19, add = FALSE)
y0                  <- matrix(c(1, 0, -1, 0, 2, 2, -2, 2, -3, -4), ncol = 2,
                              nrow = 5, byrow = TRUE)
example4.nullclines <- nullclines(example4, xlim = c(-3, 3), ylim = c(-5, 5))
example4.trajectory <- trajectory(example4, y0 = y0, tlim = c(0,10))

}
\seealso{
\code{\link{ode}}
}
\author{
Michael J. Grayling
}
