% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_hmc.R
\name{fit_hmc}
\alias{fit_hmc}
\title{HMC backend for phase models}
\usage{
fit_hmc(stan_data, chains, iter, warmup, cores, ...)
}
\arguments{
\item{stan_data}{Prepared data list}

\item{chains}{Number of chains}

\item{iter}{Iterations per chain}

\item{warmup}{Warmup iterations}

\item{cores}{Number of cores}

\item{...}{Additional HMC parameters}
}
\value{
Raw fit object
}
\description{
HMC backend for phase models
}
\keyword{internal}
