% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_hmc.R
\name{fit_laplace}
\alias{fit_laplace}
\title{Laplace approximation backend (placeholder)}
\usage{
fit_laplace(stan_data, ...)
}
\description{
Laplace approximation backend (placeholder)
}
\keyword{internal}
