% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{half_cauchy}
\alias{half_cauchy}
\title{Half-Cauchy prior distribution}
\usage{
half_cauchy(scale = 1)
}
\arguments{
\item{scale}{Scale parameter (must be positive)}
}
\value{
A prior specification object
}
\description{
A Cauchy(0, scale) distribution truncated to positive values.
Has heavier tails than half-normal.
}
\examples{
prior(sigma ~ half_cauchy(1))

}
