% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{half_normal}
\alias{half_normal}
\title{Half-Normal prior distribution}
\usage{
half_normal(sd = 1)
}
\arguments{
\item{sd}{Standard deviation of the underlying normal (must be positive)}
}
\value{
A prior specification object
}
\description{
A Normal(0, sd) distribution truncated to positive values.
Suitable for scale parameters like sigma.
}
\examples{
prior(sigma ~ half_normal(1))

}
