% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lindemannMechanism.R
\name{lindemannMechanism}
\alias{lindemannMechanism}
\title{The Lindemann Mechanism}
\usage{
lindemannMechanism(t, y, parameters)
}
\arguments{
\item{t}{Value of t, the independent variable, to evaluate the derivative
at. Should be a single number.}

\item{y}{Values of x and y, the dependent variables, to evaluate the
derivative at. Should be a vector of length 2.}

\item{parameters}{Values of the parameters of the system. Should be a number
for the value of alpha.}
}
\value{
Returns a list dy containing the values of the two derivatives at
(t, x, y).
}
\description{
The derivative function of the non-dimensional version of the Lindemann
Mechanism, an example of a two dimensional autonomous ODE system.
}
\details{
Evaluates the derivative of the following coupled ODE system at the point
(t, x, y):

dx/dt = -x^2 + alpha*x*y, dy/dt = x^2 - alpha*x*y - y.

Format is designed to be compatible with ode from the deSolve package.
}
\seealso{
\code{\link{ode}}
}
\author{
Michael J. Grayling
}
