% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{lkj}
\alias{lkj}
\title{LKJ prior for correlation matrices}
\usage{
lkj(eta = 2)
}
\arguments{
\item{eta}{Shape parameter (must be positive). eta = 1 is uniform,
larger values favor weaker correlations.}
}
\value{
A prior specification object
}
\description{
The LKJ(eta) prior places more mass on the identity matrix
as eta increases. eta = 1 is uniform over correlation matrices.
}
\examples{
prior(cor ~ lkj(2))

}
