% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_laplace.R
\name{MASS_ginv}
\alias{MASS_ginv}
\title{Generalized inverse (Moore-Penrose pseudo-inverse)}
\usage{
MASS_ginv(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{Matrix to invert}

\item{tol}{Tolerance for singular values}
}
\value{
Pseudo-inverse of X
}
\description{
Simple implementation without MASS dependency.
}
\keyword{internal}
