% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs.phase_fit}
\alias{nobs.phase_fit}
\title{Number of observations in phase_fit}
\usage{
\method{nobs}{phase_fit}(object, ...)
}
\arguments{
\item{object}{A \code{phase_fit} object}

\item{...}{Additional arguments (unused)}
}
\value{
Integer count of observations
}
\description{
Number of observations in phase_fit
}
