% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{normal}
\alias{normal}
\title{Normal prior distribution}
\usage{
normal(mean = 0, sd = 1)
}
\arguments{
\item{mean}{Mean of the distribution}

\item{sd}{Standard deviation (must be positive)}
}
\value{
A prior specification object
}
\description{
Normal prior distribution
}
\examples{
prior(beta ~ normal(0, 2.5))

}
