% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{parse_prior_specs}
\alias{parse_prior_specs}
\title{Parse prior specifications from formula syntax}
\usage{
parse_prior_specs(dots)
}
\description{
Parse prior specifications from formula syntax
}
\keyword{internal}
