% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_model.R
\name{phase_model}
\alias{phase_model}
\title{Compose a phase model}
\usage{
phase_model(phases, ...)
}
\arguments{
\item{phases}{A \code{phase_spec} object from \code{phases()}}

\item{...}{\code{phase_transition} and \code{phase_dynamics} objects}
}
\value{
A \code{phase_model} object
}
\description{
Compose a phase model
}
\examples{
model <- phase_model(
  phases("inactive", "active"),
  transition("inactive", "active", ~ treatment),
  dynamics("inactive", y ~ 1),
  dynamics("active", y ~ treatment)
)

}
