% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_phase_internal}
\alias{predict_phase_internal}
\title{Internal function for phase probability computation}
\usage{
predict_phase_internal(fit, newdata)
}
\description{
Internal function for phase probability computation
}
\keyword{internal}
