% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_response}
\alias{predict_response}
\title{Predict response marginalizing over phases}
\usage{
predict_response(fit, newdata)
}
\description{
Predict response marginalizing over phases
}
\keyword{internal}
