% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{prepare_stan_data_k}
\alias{prepare_stan_data_k}
\title{Prepare data for k-phase C++ likelihood}
\usage{
prepare_stan_data_k(
  model,
  data,
  unit_start,
  unit_end,
  n_units,
  k_phases,
  phase_names,
  response_var
)
}
\description{
Prepare data for k-phase C++ likelihood
}
\keyword{internal}
