% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{print.VarCorr_phaseR}
\alias{print.VarCorr_phaseR}
\title{Print method for VarCorr}
\usage{
\method{print}{VarCorr_phaseR}(x, ...)
}
\arguments{
\item{x}{A \code{VarCorr_phaseR} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for VarCorr
}
