% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior_to_cpp}
\alias{prior_to_cpp}
\title{Convert prior specification to C++ format}
\usage{
prior_to_cpp(prior)
}
\arguments{
\item{prior}{A phase_prior object}
}
\value{
List with type codes and parameter vectors for each parameter class
}
\description{
Creates a list that can be passed to C++ with type codes and parameters.
}
\keyword{internal}
