% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{psis_smooth}
\alias{psis_smooth}
\title{Apply Pareto smoothing to importance ratios}
\usage{
psis_smooth(log_ratios)
}
\arguments{
\item{log_ratios}{Log importance ratios}
}
\value{
List with smoothed log_weights and Pareto k estimate
}
\description{
Fits a generalized Pareto distribution to the tail of the importance
ratios and replaces extreme values with expected order statistics.
}
\keyword{internal}
