% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{ranef}
\alias{ranef}
\title{Extract random effects}
\usage{
ranef(
  object,
  component = c("all", "transition", "dynamics"),
  phase = NULL,
  summary = TRUE
)
}
\arguments{
\item{object}{A \code{phase_fit} object}

\item{component}{Which component: "transition", "dynamics", or "all" (default)}

\item{phase}{For dynamics RE, which phase (ignored for transition)}

\item{summary}{If TRUE (default), return posterior summary; if FALSE, return draws}
}
\value{
A data frame with random effect estimates
}
\description{
Extract random effects
}
\examples{
\dontrun{
ranef(fit)
ranef(fit, component = "transition")
ranef(fit, component = "dynamics", phase = "active")
}

}
