% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_laplace.R
\name{sample_mvnorm}
\alias{sample_mvnorm}
\title{Sample from multivariate normal}
\usage{
sample_mvnorm(n, mu, sigma)
}
\arguments{
\item{n}{Number of samples}

\item{mu}{Mean vector}

\item{sigma}{Covariance matrix}
}
\value{
Matrix of samples (n x length(mu))
}
\description{
Sample from multivariate normal
}
\keyword{internal}
