% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_phaseR}
\alias{sim_phaseR}
\title{Simulate data from a phase model}
\usage{
sim_phaseR(
  n_units = 100,
  n_times = 5,
  beta_trans = c(-1, 0.5),
  beta_0 = c(10, 0),
  beta_1 = c(8, -0.5),
  sigma_0 = 1,
  sigma_1 = 1.5,
  seed = NULL
)
}
\arguments{
\item{n_units}{Number of units}

\item{n_times}{Number of time points per unit}

\item{beta_trans}{Transition coefficients}

\item{beta_0}{Phase 0 dynamics coefficients}

\item{beta_1}{Phase 1 dynamics coefficients}

\item{sigma_0}{Phase 0 residual SD}

\item{sigma_1}{Phase 1 residual SD}

\item{seed}{Random seed}
}
\value{
A data frame suitable for \code{fit_phaseR()}
}
\description{
Simulate data from a phase model
}
\examples{
dat <- sim_phaseR(n_units = 100, n_times = 5)
head(dat)

}
