% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_phaseR_k}
\alias{sim_phaseR_k}
\title{Simulate data from a k-phase model}
\usage{
sim_phaseR_k(
  n_units = 100,
  n_times = 8,
  k = 3,
  beta_trans = NULL,
  beta_dyn = NULL,
  sigma = NULL,
  seed = NULL
)
}
\arguments{
\item{n_units}{Number of units}

\item{n_times}{Number of time points per unit}

\item{k}{Number of phases (must be >= 2)}

\item{beta_trans}{List of transition coefficient vectors (length k-1)}

\item{beta_dyn}{List of dynamics coefficient vectors (length k)}

\item{sigma}{Vector of residual SDs (length k)}

\item{seed}{Random seed}
}
\value{
A data frame suitable for \code{fit_phaseR()}
}
\description{
Simulate data from a k-phase model
}
\examples{
# 3-phase model
dat <- sim_phaseR_k(n_units = 50, n_times = 8, k = 3)
table(dat$true_phase)

}
