% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{smooth_pareto_tail}
\alias{smooth_pareto_tail}
\title{Smooth Pareto tail values}
\usage{
smooth_pareto_tail(x, k, m)
}
\arguments{
\item{x}{Exceedances}

\item{k}{Shape parameter}

\item{m}{Number of tail samples}
}
\value{
Smoothed exceedances
}
\description{
Replaces extreme values with expected order statistics from fitted
generalized Pareto distribution.
}
\keyword{internal}
