% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{student_t}
\alias{student_t}
\title{Student-t prior distribution}
\usage{
student_t(df = 3, mean = 0, scale = 2.5)
}
\arguments{
\item{df}{Degrees of freedom (must be positive)}

\item{mean}{Location parameter}

\item{scale}{Scale parameter (must be positive)}
}
\value{
A prior specification object
}
\description{
Student-t prior distribution
}
\examples{
# Robust prior for coefficients
prior(beta ~ student_t(3, 0, 2.5))

}
