% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_model}
\alias{validate_model}
\title{Validate phase model specification}
\usage{
validate_model(model)
}
\arguments{
\item{model}{A \code{phase_model} object}
}
\value{
Invisibly returns TRUE if valid, otherwise stops with error
}
\description{
Validate phase model specification
}
\keyword{internal}
