% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_model_data}
\alias{validate_model_data}
\title{Validate phase model against data}
\usage{
validate_model_data(model, data)
}
\arguments{
\item{model}{A \code{phase_model} object}

\item{data}{A data frame}
}
\value{
Invisibly returns TRUE if valid, otherwise stops with error
}
\description{
Validate phase model against data
}
\keyword{internal}
