% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{validate_nesting}
\alias{validate_nesting}
\title{Validate nested structure}
\usage{
validate_nesting(data, outer, inner)
}
\arguments{
\item{data}{Data frame}

\item{outer}{Outer grouping variable name}

\item{inner}{Inner grouping variable name}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
Checks that nested groups are properly nested (each level of b is within one level of a).
}
\keyword{internal}
