% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{validate_prior_for_class}
\alias{validate_prior_for_class}
\title{Validate that a prior distribution is appropriate for a parameter class}
\usage{
validate_prior_for_class(param_class, prior_obj)
}
\description{
Validate that a prior distribution is appropriate for a parameter class
}
\keyword{internal}
