% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{VarCorr}
\alias{VarCorr}
\title{Extract variance components}
\usage{
VarCorr(object)
}
\arguments{
\item{object}{A \code{phase_fit} object}
}
\value{
A data frame with variance component estimates
}
\description{
Extract variance components
}
