% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseR.R
\name{.paramDummy}
\alias{.paramDummy}
\title{A function such that we can apply DRY in param documentation}
\usage{
.paramDummy(state.names)
}
\arguments{
\item{state.names}{State names for ode functions that do not use positional
states}
}
\description{
A function such that we can apply DRY in param documentation
}
