% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotkaVolterra.R
\name{lotkaVolterra}
\alias{lotkaVolterra}
\title{The Lotka-Volterra model}
\usage{
lotkaVolterra(t, y, parameters)
}
\arguments{
\item{t}{The value of \ifelse{html}{\out{<i>t</i>}}{\eqn{t}}, the independent
variable, to evaluate the derivative at. Should be a
\code{\link[base]{numeric}} \code{\link[base]{vector}} of
\code{\link[base]{length}} one.}

\item{y}{The values of \ifelse{html}{\out{<i>x</i>}}{\eqn{x}} and
\ifelse{html}{\out{<i>y</i>}}{\eqn{y}}, the dependent variables, to evaluate
the derivative at. Should be a \code{\link[base]{numeric}}
\code{\link[base]{vector}} of \code{\link[base]{length}} two.}

\item{parameters}{The values of the parameters of the system. Should be a
\code{\link[base]{numeric}} \code{\link[base]{vector}} with parameters
specified in the following order:
\ifelse{html}{\out{<i>&lambda;</i>}}{\eqn{\lambda}},
\ifelse{html}{\out{<i>&epsilon;</i>}}{\eqn{\epsilon}},
\ifelse{html}{\out{<i>&eta;</i>}}{\eqn{\eta}},
\ifelse{html}{\out{<i>&delta;</i>}}{\eqn{\delta}}.}
}
\value{
Returns a \code{\link[base]{list}} containing the values of the two
derivatives at
\ifelse{html}{\out{(<i>t</i>, <i>x</i>, <i>y</i>)}}{\eqn{(t, x, y)}}.
}
\description{
The derivative function of the Lotka-Volterra model, an example of a
two-dimensional autonomous ODE system.
}
\details{
\code{lotkaVolterra} evaluates the derivative of the following ODE at the
point \ifelse{html}{\out{(<i>t</i>, <i>x</i>, <i>y</i>)}}{\eqn{(t, x, y)}}:

\ifelse{html}{\out{<i>dx</i>/<i>dt</i> = <i>&lambda;x</i> -
<i>&epsilon;xy</i>, <i>dy</i>/<i>dt</i> = <i>&eta;xy</i> -
<i>&delta;y</i>.}}{\deqn{\frac{dx}{dt} = \lambda x - \epsilon xy,
\frac{dy}{dt} = \eta xy - \delta y.}}

Its format is designed to be compatible with \code{\link[deSolve]{ode}} from
the \code{\link[deSolve]{deSolve}} package.
}
\seealso{
\code{\link[deSolve]{ode}}
}
\author{
Michael J Grayling
}
