% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{parse_formula_re}
\alias{parse_formula_re}
\title{Parse formula for random effects}
\usage{
parse_formula_re(formula)
}
\arguments{
\item{formula}{A formula potentially containing \code{(1 | group)} terms}
}
\value{
A list with fixed formula, RE terms, and group names
}
\description{
Parse formula for random effects
}
\keyword{internal}
